added view for controllers (setting and permission)
